-Penwing - relacion entre entidades y creacion de usuario -creacion de entidad Credit/Frontend -validacion Login _walley e informacion -Correccion a BigDecimal. -BUy/Sell -mensaje de recepcion y DTO

+++ src/main/java/com/Testing/practicasTesteo/entity/Credit.java
@@ -25,12 +25,12 @@ public class Credit {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long creditId;
 
-    private BigDecimal euro;
-    private BigDecimal bitcoin;
-    private BigDecimal ethereum;
-    private BigDecimal ripple;
-    private BigDecimal litecoin;
-    private BigDecimal cardano;
+    private Double euro;
+    private Double bitcoin;
+    private Double ethereum;
+    private Double ripple;
+    private Double litecoin;
+    private Double cardano;
 
     @ManyToOne
     @JoinColumn(name = "customer_id", referencedColumnName = "customerId")
